services/call/scripts/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.9.0
twilio==9.8.5
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
yarl==1.22.0
